+++ internal/console/cmd/spire_init_cmd.go
@@ -2,8 +2,10 @@ package cmd
 
 import (
 	"github.com/Akkadius/spire/internal/spire"
+	"github.com/shirou/gopsutil/v3/mem"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
+	"runtime"
 )
 
 type SpireInitCommand struct {
@@ -18,6 +20,8 @@ func (c *SpireInitCommand) Command() *cobra.Command {
 
 // flagAuthEnabled is a flag that can be passed to the spire:init command
 var flagAuthEnabled bool
+var compileServer bool
+var compileLocation string
 
 // NewSpireInitCommand creates a new spire:init command
 func NewSpireInitCommand(
@@ -34,7 +38,10 @@ func NewSpireInitCommand(
 		},
 	}
 
+	// TODO: Change these later
 	i.command.PersistentFlags().BoolVarP(&flagAuthEnabled, "auth-enabled", "a", true, "Whether or not Spire has authentication / authorization enabled")
+	i.command.PersistentFlags().BoolVarP(&compileServer, "compile-server", "c", false, "Whether or not the server should be compiled (default: false - uses releases)")
+	i.command.PersistentFlags().StringVarP(&compileLocation, "compile-build-location", "l", "~/code/build", "Determines where the binaries should be built")
 	i.command.Run = i.Handle
 
 	return i
@@ -50,13 +57,34 @@ func (c *SpireInitCommand) Handle(_ *cobra.Command, args []string) {
 	if flagAuthEnabled {
 		authEnabled = 1
 	}
+	cores := 1
+
+	// get system memory available
+	memory, err := mem.VirtualMemory()
+	if err != nil {
+		c.logger.Fatal(err)
+	}
+
+	// get system memory available in GB
+	memoryAvailableGb := memory.Available / 1024 / 1024 / 1024
+	if memoryAvailableGb >= 10 {
+		cores = runtime.NumCPU() - 4
+		if cores < 1 {
+			cores = 1
+		}
+	} else if memoryAvailableGb >= 6 {
+		cores = 4
+	}
 
 	// Run the spire:init command
-	err := c.spireinit.InitApp(
+	err = c.spireinit.InitApp(
 		&spire.InitAppRequest{
-			AuthEnabled: authEnabled,
-			Username:    username,
-			Password:    password,
+			AuthEnabled:   authEnabled,
+			Username:      username,
+			Password:      password,
+			SelfCompiled:  compileServer,
+			BuildLocation: compileLocation,
+			BuildCores:    cores,
 		},
 	)
 	if err != nil {